Split out chat filtering logic into separate class for testing
